Fixed compile error when B3_USE_DOUBLE_PRECISION is defined.
When `B3_USE_DOUBLE_PRECISION` is defined, B3_USE_SSE is disabled, and b3Scalar is a typedef of double. However, m_floats is still a float, which causes a compile error with `b3Vector3::operator b3Scalar*()` and `b3Vector3::operator const b3Scalar*() const`  found on lines 536 and 537.